# gpao-tex-ui/source/main/src/openapi/product/openapi.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "OpenApi specification - Oussama Soussi",
      "description": "OpenApi documentation for product service",
      "contact": {
        "name": "Oussama",
        "email": "oussama.soussi.it@gmail.com"
      },
      "version": "1.0"
    },
    "servers": [
      {
        "url": "http://localhost:2323/produit/",
        "description": "Product API"
      }
    ],
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
...
      "/api/produit/couleur/creerCouleur": {
        "post": {
          "tags": [
            "couleur-controller"
          ],
          "operationId": "createCouleur",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CouleurValue"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/Couleur"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/produit/couleur/find/by/{id}": {
        "get": {
          "tags": [
            "couleur-controller"
          ],
          "operationId": "findCouleurById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/CouleurValue"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/produit/couleur/delete/{id}": {
        "delete": {
          "tags": [
            "couleur-controller"
          ],
          "operationId": "deleteCouleur",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "Couleur": {
          "type": "object",
          "properties": {
            "creationDate": {
              "type": "string",
              "format": "date-time"
            },
            "modificationDate": {
              "type": "string",
              "format": "date-time"
            },
            "deletionDate": {
              "type": "string",
              "format": "date-time"
            },
            "creatorIdentifier": {
              "type": "string"
            },
            "latestUpdaterId": {
              "type": "string"
            },
            "version": {
              "type": "integer",
              "format": "int64"
            },
            "archived": {
              "type": "boolean"
            },
            "produit": {
              "$ref": "#/components/schemas/Produit"
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "designation": {
              "type": "string"
            },
            "codeCouleur": {
              "type": "string"
            }
          }
        },
        "ElementTraitement": {
          "type": "object",
          "properties": {
            "creationDate": {
              "type": "string",
              "format": "date-time"
            },
            "modificationDate": {
              "type": "string",
              "format": "date-time"
            },
            "deletionDate": {
              "type": "string",
              "format": "date-time"
            },
            "creatorIdentifier": {
              "type": "string"
            },
            "latestUpdaterId": {
              "type": "string"
            },
            "version": {
              "type": "integer",
              "format": "int64"
            },
            "archived": {
              "type": "boolean"
            },
            "produit": {
              "$ref": "#/components/schemas/Produit"
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "dimension": {
              "type": "string"
            },
            "observation": {
              "type": "string"
            },
            "machineId": {
              "type": "integer",
              "format": "int64"
            },
            "code": {
              "type": "string"
            },
            "nature": {
              "type": "string"
            },
            "mesure": {
              "type": "number",
              "format": "double"
            },
            "tolerancePlus": {
              "type": "number",
              "format": "double"
            },
            "toleranceMoins": {
              "type": "number",
              "format": "double"
            }
          }
        },
        "FamilleProduit": {
          "type": "object",
          "properties": {
            "creationDate": {
              "type": "string",
              "format": "date-time"
            },
            "modificationDate": {
              "type": "string",
              "format": "date-time"
            },
            "deletionDate": {
              "type": "string",
              "format": "date-time"
            },
            "creatorIdentifier": {
              "type": "string"
            },
            "latestUpdaterId": {
              "type": "string"
            },
            "version": {
              "type": "integer",
              "format": "int64"
            },
            "archived": {
              "type": "boolean"
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "designation": {
              "type": "string"
            },
            "achat": {
              "type": "boolean"
            },
            "vente": {
              "type": "boolean"
            }
          }
        },
        "LigneProduit": {
          "type": "object",
          "properties": {
            "creationDate": {
              "type": "string",
              "format": "date-time"
            },
            "modificationDate": {
              "type": "string",
              "format": "date-time"
            },
            "deletionDate": {
              "type": "string",
              "format": "date-time"
            },
            "creatorIdentifier": {
              "type": "string"
            },
            "latestUpdaterId": {
              "type": "string"
            },
            "version": {
              "type": "integer",
              "format": "int64"
            },
            "archived": {
              "type": "boolean"
            },
            "theme": {
              "$ref": "#/components/schemas/Theme"
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "designation": {
              "type": "string"
            },
            "fournisseur": {
              "type": "integer",
              "format": "int64"
            },
            "finDe": {
              "type": "string",
              "format": "date-time"
            },
            "finA": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "MesureProduit": {
          "type": "object",
          "properties": {
            "creationDate": {
              "type": "string",
              "format": "date-time"
            },
            "modificationDate": {
              "type": "string",
              "format": "date-time"
            },
            "deletionDate": {
              "type": "string",
              "format": "date-time"
            },
            "creatorIdentifier": {
              "type": "string"
            },
            "latestUpdaterId": {
              "type": "string"
            },
            "version": {
              "type": "integer",
              "format": "int64"
            },
            "archived": {
              "type": "boolean"
            },
            "produit": {
              "$ref": "#/components/schemas/Produit"
            },
            "taille": {
              "$ref": "#/components/schemas/Taille"
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "code": {
              "type": "string"
            },
            "mesure": {
              "type": "number",
              "format": "double"
            },
            "tolerancePlus": {
              "type": "number",
              "format": "double"
            },
            "toleranceMoins": {
              "type": "number",
              "format": "double"
            }
          }
        },
        "PilotageEfficience": {
          "type": "object",
          "properties": {
            "creationDate": {
              "type": "string",
              "format": "date-time"
            },
            "modificationDate": {
              "type": "string",
              "format": "date-time"
            },
            "deletionDate": {
              "type": "string",
              "format": "date-time"
            },
            "creatorIdentifier": {
              "type": "string"
            },
            "latestUpdaterId": {
              "type": "string"
            },
            "version": {
              "type": "integer",
              "format": "int64"
            },
            "archived": {
              "type": "boolean"
            },
            "site": {
              "$ref": "#/components/schemas/Site"
            },
            "produit": {
              "$ref": "#/components/schemas/Produit"
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "efficienceJour1": {
              "type": "number",
              "format": "double"
            },
            "efficienceJour2": {
              "type": "number",
              "format": "double"
            },
            "efficienceJour3": {
              "type": "number",
              "format": "double"
            },
            "efficienceJour4": {
              "type": "number",
              "format": "double"
            },
            "efficienceJour5": {
              "type": "number",
              "format": "double"
            },
            "efficienceJour6": {
              "type": "number",
              "format": "double"
            },
            "efficienceJour7": {
              "type": "number",
              "format": "double"
            }
          }
        },
        "Produit": {
          "type": "object",
          "properties": {
            "creationDate": {
              "type": "string",
              "format": "date-time"
            },
            "modificationDate": {
              "type": "string",
              "format": "date-time"
            },
            "deletionDate": {
              "type": "string",
              "format": "date-time"
            },
            "creatorIdentifier": {
              "type": "string"
            },
            "latestUpdaterId": {
              "type": "string"
            },
            "version": {
              "type": "integer",
              "format": "int64"
            },
            "archived": {
              "type": "boolean"
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "sousFamilleProduit": {
              "$ref": "#/components/schemas/SousFamilleProduit"
            },
            "site": {
              "$ref": "#/components/schemas/Site"
            },
            "ligneProduit": {
              "$ref": "#/components/schemas/LigneProduit"
            },
            "produitGoShippements": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ProduitGoShippement"
              }
            },
            "mesureProduits": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/MesureProduit"
              }
            },
            "tailles": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Taille"
              }
            },
            "couleurs": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Couleur"
              }
            },
            "pilotageEfficiences": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/PilotageEfficience"
              }
            },
            "elementTraitements": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ElementTraitement"
              }
            },
            "partieInteresseId": {
              "type": "integer",
              "format": "int64"
            },
            "reference": {
              "type": "string"
            },
            "designation": {
              "type": "string"
            },
            "ficheBesoin": {
              "type": "boolean"
            },
            "sansTaille": {
              "type": "boolean"
            },
            "sansCouleur": {
              "type": "boolean"
            },
            "codeConception": {
              "type": "string"
            },
            "quantite": {
              "type": "integer",
              "format": "int32"
            },
            "cadence": {
              "type": "integer",
              "format": "int32"
            },
            "prixUnitaire": {
              "type": "number"
            },
            "prixMajore": {
              "type": "number"
            },
            "etat": {
              "type": "string"
            }
          }
        },
        "ProduitGoShippement": {
          "type": "object",
          "properties": {
            "creationDate": {
              "type": "string",
              "format": "date-time"
            },
            "modificationDate": {
              "type": "string",
              "format": "date-time"
            },
            "deletionDate": {
              "type": "string",
              "format": "date-time"
            },
            "creatorIdentifier": {
              "type": "string"
            },
            "latestUpdaterId": {
              "type": "string"
            },
            "version": {
              "type": "integer",
              "format": "int64"
            },
            "archived": {
              "type": "boolean"
            },
            "site": {
              "$ref": "#/components/schemas/Site"
            },
            "produit": {
              "$ref": "#/components/schemas/Produit"
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "saison": {
              "type": "string"
            },
            "goShippement": {
              "type": "string"
            }
          }
        },
        "Site": {
          "type": "object",
          "properties": {
            "creationDate": {
              "type": "string",
              "format": "date-time"
            },
            "modificationDate": {
              "type": "string",
              "format": "date-time"
            },
            "deletionDate": {
              "type": "string",
              "format": "date-time"
            },
            "creatorIdentifier": {
              "type": "string"
            },
            "latestUpdaterId": {
              "type": "string"
            },
            "version": {
              "type": "integer",
              "format": "int64"
            },
            "archived": {
              "type": "boolean"
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "nom": {
              "type": "string"
            },
            "designation": {
              "type": "string"
            },
            "abriviation": {
              "type": "string"
            },
            "sitePrincipal": {
              "type": "boolean"
            },
            "effidapSiteId": {
              "type": "integer",
              "format": "int64"
            },
            "production": {
              "type": "boolean"
            },
            "debut": {
              "type": "integer",
              "format": "int64"
            },
            "fin1": {
              "type": "integer",
              "format": "int64"
            },
            "fin2": {
              "type": "integer",
              "format": "int64"
            },
            "fin3": {
              "type": "integer",
              "format": "int64"
            },
            "fin4": {
              "type": "integer",
              "format": "int64"
            },
            "fin5": {
              "type": "integer",
              "format": "int64"
            }
          }
        },
        "SousFamilleProduit": {
          "type": "object",
          "properties": {
            "creationDate": {
              "type": "string",
              "format": "date-time"
            },
            "modificationDate": {
              "type": "string",
              "format": "date-time"
            },
            "deletionDate": {
              "type": "string",
              "format": "date-time"
            },
            "creatorIdentifier": {
              "type": "string"
            },
            "latestUpdaterId": {
              "type": "string"
            },
            "version": {
              "type": "integer",
              "format": "int64"
            },
            "archived": {
              "type": "boolean"
            },
            "familleProduit": {
              "$ref": "#/components/schemas/FamilleProduit"
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "designation": {
              "type": "string"
            }
          }
        },
        "Taille": {
          "type": "object",
          "properties": {
            "creationDate": {
              "type": "string",
              "format": "date-time"
            },
            "modificationDate": {
              "type": "string",
              "format": "date-time"
            },
            "deletionDate": {
              "type": "string",
              "format": "date-time"
            },
            "creatorIdentifier": {
              "type": "string"
            },
            "latestUpdaterId": {
              "type": "string"
            },
            "version": {
              "type": "integer",
              "format": "int64"
            },
            "archived": {
              "type": "boolean"
            },
            "produit": {
              "$ref": "#/components/schemas/Produit"
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "designation": {
              "type": "string"
            },
            "ordre": {
              "type": "integer",
              "format": "int32"
            },
            "code": {
              "type": "string"
            }
          }
        },
        "Theme": {
          "type": "object",
          "properties": {
            "creationDate": {
              "type": "string",
              "format": "date-time"
            },
            "modificationDate": {
              "type": "string",
              "format": "date-time"
            },
            "deletionDate": {
              "type": "string",
              "format": "date-time"
            },
            "creatorIdentifier": {
              "type": "string"
            },
            "latestUpdaterId": {
              "type": "string"
            },
            "version": {
              "type": "integer",
              "format": "int64"
            },
            "archived": {
              "type": "boolean"
            },
            "produit": {
              "$ref": "#/components/schemas/Produit"
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "designation": {
              "type": "string"
            },
            "spl": {
              "type": "string"
            }
          }
        },
...
        "CouleurValue": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "produitId": {
              "type": "integer",
              "format": "int64"
            },
            "designation": {
              "type": "string"
            },
            "codeCouleur": {
              "type": "string"
            }
          }
        },
...
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "description": "JWT auth descirption",
          "in": "header",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }